Renaming module 'scipion4facilities' to 'emfacilities'
