+++ emfacilities/protocols/protocol_monitor_movie_gain.py
