+++ requirements.txt
@@ -1,4 +1,3 @@
-scipion-pyworkflow
 scipion-em
 nvidia-ml-py3
 pytz
